+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/studio-stomp/craft-twig-linter/compare/0.4.0...main)
+## [Unreleased](https://github.com/studio-stomp/craft-twig-linter/compare/0.4.1...develop)
+
+## [0.4.1](https://github.com/studio-stomp/craft-twig-linter/compare/0.4.0...0.4.1) - 2024-08-15
+### Changed
+- feat: upgrade tests to Codeception 5
 
 ## [0.4.0](https://github.com/studio-stomp/craft-twig-linter/compare/0.3.1...0.4.0) - 2023-03-02
 ### Changed